update bridge integration tests
