+++ integrationTests/chainSimulator/bridge/common.go
@@ -18,7 +18,6 @@ import (
 
 const (
 	issuePaymentCost         = "50000000000000000" // esdt-safe contract without header-verifier checks
-	esdtSafeWasmPath         = "testdata/mvx-esdt-safe.wasm"
 	enshrineEsdtSafeWasmPath = "testdata/enshrine-esdt-safe.wasm"
 	//enshrine esdt-safe contract without checks for prefix or issue cost paid for new tokens
 	simpleEnshrineEsdtSafeWasmPath = "testdata/simple-enshrine-esdt-safe.wasm"
@@ -116,19 +115,6 @@ func deployBridgeSetup(
 			Nonce:  nonce,
 		},
 	}
-}
-
-func esdtSafeContract(
-	t *testing.T,
-	cs chainSim.ChainSimulator,
-	ownerAddress []byte,
-	nonce *uint64,
-	systemContractDeploy []byte,
-	_ ArgsEsdtSafe,
-	contractWasmPath string,
-) []byte {
-	esdtSafeArgs := "@00000000000000000500f3c3c6f64f8a20ced95a45a7de596ab08f9082d0f8ef" // dummy header_verifier_address
-	return chainSim.DeployContract(t, cs, ownerAddress, nonce, systemContractDeploy, esdtSafeArgs, contractWasmPath)
 }
 
 func enshrineEsdtSafeContract(
